+++ packages/utils/src/health.rs
@@ -1,34 +1,27 @@
 use alloy_primitives::U256;
 use alloy_provider::Provider;
 use alloy_rpc_types_eth::SyncStatus;
-use anyhow::Context;
 use thiserror::Error;
 
 use crate::{
-    config::{ChainConfigs, CosmosChainConfig, EvmChainConfig, EvmChainConfigExt},
+    config::{CosmosChainConfig, EvmChainConfig, EvmChainConfigExt},
     error::EvmClientError,
     evm_client::EvmQueryClient,
 };
 use wavs_types::{AnyChainConfig, ChainKey};
 
-pub async fn health_check_chains_query(
-    configs: &ChainConfigs,
-    keys: &[ChainKey],
+pub async fn health_check_single_chain(
+    key: &ChainKey,
+    config: &AnyChainConfig,
 ) -> anyhow::Result<()> {
-    for key in keys {
-        let config = configs
-            .get_chain(key)
-            .context(format!("Failed to get config for chain key: {key}"))?;
-
-        match config {
-            AnyChainConfig::Evm(config) => {
-                check_evm_chain_health_query(key.clone(), config).await?;
-                tracing::info!("Evm chain [{key}] is healthy");
-            }
-            AnyChainConfig::Cosmos(config) => {
-                check_cosmos_chain_health_query(key.clone(), config).await?;
-                tracing::info!("Cosmos chain [{key}] is healthy");
-            }
+    match config {
+        AnyChainConfig::Evm(config) => {
+            check_evm_chain_health_query(key.clone(), config.clone()).await?;
+            tracing::info!("Evm chain [{key}] is healthy");
+        }
+        AnyChainConfig::Cosmos(config) => {
+            check_cosmos_chain_health_query(key.clone(), config.clone()).await?;
+            tracing::info!("Cosmos chain [{key}] is healthy");
         }
     }
     Ok(())